TestPostman

+++ gestion-biblioteca/src/main/java/com/uisrael/gestion_biblioteca/service/GeneroService.java
@@ -8,6 +8,6 @@ public interface GeneroService {
     List<Genero> findAllGenero();
     Genero saveGenero(Genero genero);
     void deleteGenero(int id);
-    Genero updateGenero(Long id, Genero detalles);
+    Genero updateGenero(int id, Genero detalles);
 
 }
